Refactor : goods 컨벤션 맞춰 수정 - SignUpDto로 묶음
